# ExpenseManagement/ExpenseManagement/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExpenseManagementAPI",
    "version": "v1"
  },
  "paths": {
...
    "/api/Income/createIncome": {
      "post": {
        "tags": [
          "Income"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              }
            }
          }
        }
      }
    },
    "/api/Income/getAllIncomes/{userId}": {
      "get": {
        "tags": [
          "Income"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncomeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Income/getIncome/{id}": {
      "get": {
        "tags": [
          "Income"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Income/editIncome/{id}": {
      "put": {
        "tags": [
          "Income"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IncomeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Income"
                }
              }
            }
          }
        }
      }
    },
    "/api/Income/deleteIncome/{id}": {
      "delete": {
        "tags": [
          "Income"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RP/createRP": {
      "post": {
        "tags": [
          "RP"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              }
            }
          }
        }
      }
    },
    "/api/RP/getAllRPs/{userId}": {
      "get": {
        "tags": [
          "RP"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularPaymentResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularPaymentResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegularPaymentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/RP/getRP/{id}": {
      "get": {
        "tags": [
          "RP"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPaymentResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPaymentResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPaymentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/RP/updateRP/{id}": {
      "put": {
        "tags": [
          "RP"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegularPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegularPayment"
                }
              }
            }
          }
        }
      }
    },
    "/api/RP/deleteRP/{id}": {
      "delete": {
        "tags": [
          "RP"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ClientOptions": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "string",
            "nullable": true
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "queryParams": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Income": {
        "type": "object",
        "properties": {
          "baseUrl": {
            "type": "string",
            "nullable": true
          },
          "requestClientOptions": {
            "$ref": "#/components/schemas/ClientOptions"
          },
          "tableName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "primaryKey": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IncomeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IncomeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegularPayment": {
        "type": "object",
        "properties": {
          "baseUrl": {
            "type": "string",
            "nullable": true
          },
          "requestClientOptions": {
            "$ref": "#/components/schemas/ClientOptions"
          },
          "tableName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "primaryKey": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true,
            "readOnly": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "recurrenceDay": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegularPaymentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "recurrenceDay": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegularPaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "float"
          },
          "recurrenceDay": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Inserisci 'Bearer' [spazio] e poi il token JWT.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}